+++ src/views/SliderView.vue
@@ -2,31 +2,57 @@
 import { ref, h, watchEffect } from 'vue'
 import { FireFilled } from '@ant-design/icons-vue'
 const singleValue = ref<number>(20)
-const smallSingleValue = ref<number>(0.5)
-const singleCustomValue = ref<number>(0)
 const doubleValue = ref<number[]>([20, 80])
-const doubleCustomValue = ref<number[]>([-5, 5])
+const disabledSingleValue = ref<number>(20)
+const disabledDoubleValue = ref<number[]>([20, 80])
+const customScaleSingleValue = ref<number>(0)
+const customScaleDoubleValue = ref<number[]>([-5, 5])
+const customStepSingleValue = ref<number>(30)
+const customStepDoubleValue = ref<number[]>([30, 60])
 const markSingleValue1 = ref<number>(24)
 const markSingleValue2 = ref<number>(30)
 const markSingleValue3 = ref<number>(37)
-const markDoubleValue1 = ref<number[]>([20, 55])
+const markDoubleValue1 = ref<number[]>([25, 75])
 const markDoubleValue2 = ref<number[]>([30, 60])
 const markDoubleValue3 = ref<number[]>([0, 37])
+const verticalSingleValue = ref<number>(20)
+const verticalDoubleValue = ref<number[]>([20, 80])
+const markVerticalSingleValue = ref<number>(30)
+const markVerticalDoubleValue1 = ref<number[]>([35, 65])
+const markVerticalDoubleValue2 = ref<number[]>([30, 60])
+const markVerticalDoubleValue3 = ref<number[]>([0, 37])
 const marks = ref<Record<number, any>>({
   0: '0°C',
   26: '26°C',
   37: '37°C',
   100: {
     style: {
-      color: '#f50',
+      color: '#f50'
     },
     label: '100°C'
-    // label: h(FireFilled)
   }
 })
-const singleCustomStepValue = ref<number>(30)
-const doubleCustomStepValue = ref<number[]>([30, 60])
-const customStyle = {
+const verticalMarks = ref<Record<number, any>>({
+  0: '0°C',
+  26: '26°C',
+  37: '37°C',
+  100: {
+    style: {
+      color: '#f50',
+      fontWeight: 'bold'
+    },
+    label: h(FireFilled)
+  }
+})
+const inputSingleValue = ref<number>(20)
+const smallSingleValue = ref<number>(0.5)
+const formatSingleValue = ref<number>(20)
+const formatDoubleValue = ref<number[]>([20, 80])
+const hideTooltipSingleValue = ref<number>(20)
+const hideTooltipDoubleValue = ref<number[]>([20, 80])
+const customStyleSingleValue = ref<number>(20)
+const customStyleDoubleValue = ref<number[]>([20, 80])
+const singleCustomStyle = {
   '--track-color': '#ffbb96',
   '--track-color-hover': '#d4380d',
   '--handle-color': '#fff2e8',
@@ -40,23 +66,29 @@ const rangeCustomStyle = {
   '--track-color-hover': '#d4380d',
   '--handle-color': '#fff2e8',
   '--handle-shadow-color': '#ffbb96',
-  '--handle-shadow-color-hover-focus': '#d4380d'
+  '--handle-shadow-color-hover-focus': '#d4380d',
+  '--dot-border-color': 'rgb(219, 219, 223)',
+  '--dot-border-color-hover': 'rgb(199, 199, 203)',
+  '--dot-color-active': '#ffbb96'
 }
-// watchEffect(() => {
-//   console.log('singleValue', singleValue.value)
-// })
-// watchEffect(() => {
-//   console.log('smallSingleValue', smallSingleValue.value)
-// })
-// watchEffect(() => {
-//   console.log('singleCustomValue', singleCustomValue.value)
-// })
-// watchEffect(() => {
-//   console.log('doubleValue', doubleValue.value)
-// })
-// watchEffect(() => {
-//   console.log('doubleCustomValue', doubleCustomValue.value)
-// })
+watchEffect(() => {
+  console.log('singleValue', singleValue.value)
+})
+watchEffect(() => {
+  console.log('doubleValue', doubleValue.value)
+})
+watchEffect(() => {
+  console.log('customScaleSingleValue', customScaleSingleValue.value)
+})
+watchEffect(() => {
+  console.log('customScaleDoubleValue', customScaleDoubleValue.value)
+})
+watchEffect(() => {
+  console.log('customStepSingleValue', customStepSingleValue.value)
+})
+watchEffect(() => {
+  console.log('customStepDoubleValue', customStepDoubleValue.value)
+})
 watchEffect(() => {
   console.log('markSingleValue1', markSingleValue1.value)
 })
@@ -75,37 +107,77 @@ watchEffect(() => {
 watchEffect(() => {
   console.log('markDoubleValue3', markDoubleValue3.value)
 })
-// watchEffect(() => {
-//   console.log('singleCustomStepValue', singleCustomStepValue.value)
-// })
-// watchEffect(() => {
-//   console.log('doubleCustomStepValue', doubleCustomStepValue.value)
-// })
+watchEffect(() => {
+  console.log('verticalSingleValue', verticalSingleValue.value)
+})
+watchEffect(() => {
+  console.log('verticalDoubleValue', verticalDoubleValue.value)
+})
+watchEffect(() => {
+  console.log('markVerticalSingleValue', markVerticalSingleValue.value)
+})
+watchEffect(() => {
+  console.log('markVerticalDoubleValue1', markVerticalDoubleValue1.value)
+})
+watchEffect(() => {
+  console.log('markVerticalDoubleValue2', markVerticalDoubleValue2.value)
+})
+watchEffect(() => {
+  console.log('markVerticalDoubleValue3', markVerticalDoubleValue3.value)
+})
+watchEffect(() => {
+  console.log('inputSingleValue', inputSingleValue.value)
+})
+watchEffect(() => {
+  console.log('smallSingleValue', smallSingleValue.value)
+})
+watchEffect(() => {
+  console.log('formatSingleValue', formatSingleValue.value)
+})
+watchEffect(() => {
+  console.log('formatDoubleValue', formatDoubleValue.value)
+})
+watchEffect(() => {
+  console.log('hideTooltipSingleValue', hideTooltipSingleValue.value)
+})
+watchEffect(() => {
+  console.log('hideTooltipDoubleValue', hideTooltipDoubleValue.value)
+})
+watchEffect(() => {
+  console.log('customStyleSingleValue', customStyleSingleValue.value)
+})
+watchEffect(() => {
+  console.log('customStyleDoubleValue', customStyleDoubleValue.value)
+})
 function onChange(value: number | number[]) {
   console.log('change', value)
 }
 function formatter(value: number) {
   return `${value}%`
 }
-const markSingleValue4 = ref(32)
 </script>
 <template>
   <div style="width: 80%">
-    <!-- <h1>{{ $route.name }} {{ $route.meta.title }}</h1>
+    <h1>{{ $route.name }} {{ $route.meta.title }}</h1>
     <h2 class="mt30 mb10">基本使用</h2>
     <Slider v-model:value="singleValue" @change="onChange" />
     <h2 class="mt30 mb10">双滑块</h2>
     <Slider range v-model:value="doubleValue" @change="onChange" />
     <h2 class="mt30 mb10">禁用</h2>
     <Flex vertical gap="large">
-      <Slider v-model:value="singleValue" disabled />
-      <Slider v-model:value="doubleValue" range disabled />
+      <Slider v-model:value="disabledSingleValue" disabled />
+      <Slider v-model:value="disabledDoubleValue" range disabled />
     </Flex>
     <h2 class="mt30 mb10">自定义数值范围</h2>
     <Flex vertical gap="large">
-      <Slider :min="-10" :max="10" v-model:value="singleCustomValue" />
-      <Slider :min="-10" :max="10" range v-model:value="doubleCustomValue" />
-    </Flex> -->
+      <Slider :min="-10" :max="10" v-model:value="customScaleSingleValue" />
+      <Slider :min="-10" :max="10" range v-model:value="customScaleDoubleValue" />
+    </Flex>
+    <h2 class="mt30 mb10">自定义步长</h2>
+    <Flex vertical gap="large">
+      <Slider :step="5" v-model:value="customStepSingleValue" />
+      <Slider range :step="5" v-model:value="customStepDoubleValue" />
+    </Flex>
     <h2 class="mt30 mb10">自定义刻度标记</h2>
     <Flex vertical>
       <Slider v-model:value="markSingleValue1" :marks="marks">
@@ -124,7 +196,6 @@ const markSingleValue4 = ref(32)
           <template v-else>{{ label }}</template>
         </template>
       </Slider>
-      <n-slider v-model:value="markSingleValue4" :marks="marks" :step="10" />
       <Slider v-model:value="markSingleValue2" :marks="marks" :step="10">
         <template #mark="{ label, value }">
           <template v-if="value === 100">
@@ -158,25 +229,27 @@ const markSingleValue4 = ref(32)
         </template>
       </Slider>
     </Flex>
-    <!-- <h2 class="mt30 mb10">自定义步长</h2>
-    <Flex vertical gap="large">
-      <Slider :step="5" v-model:value="singleCustomStepValue" />
-      <Slider range :step="5" v-model:value="doubleCustomStepValue" />
-    </Flex> -->
     <h2 class="mt30 mb10">垂直模式</h2>
     <Space :gap="36" style="height: 300px">
-      <!-- <Slider vertical v-model:value="singleValue" />
-      <Slider range vertical v-model:value="doubleValue" /> -->
-      <!-- <Slider vertical v-model:value="markSingleValue" :marks="marks" />
-      <a-slider vertical v-model:value="markSingleValue" :marks="marks" /> -->
+      <Slider vertical v-model:value="verticalSingleValue" />
+      <Slider range vertical v-model:value="verticalDoubleValue" />
+      <Slider vertical v-model:value="markVerticalSingleValue" :marks="marks" />
+      <Slider vertical range v-model:value="markVerticalDoubleValue1" :marks="marks" />
+      <Slider vertical range v-model:value="markVerticalDoubleValue2" :marks="verticalMarks" :step="10" />
+      <Slider vertical range v-model:value="markVerticalDoubleValue3" :marks="verticalMarks" step="mark">
+        <template #mark="{ label, isVNode, value }">
+          <template v-if="isVNode"> {{ value }}°C </template>
+          <template v-else>{{ label }}</template>
+        </template>
+      </Slider>
     </Space>
-    <!-- <h2 class="mt30 mb10">带输入框的滑块</h2>
+    <h2 class="mt30 mb10">带输入框的滑块</h2>
     <Row :gutter="[24, 16]">
       <Col :span="18">
-        <Slider v-model:value="singleValue" />
+        <Slider v-model:value="inputSingleValue" />
       </Col>
       <Col :span="6">
-        <InputNumber v-model:value="singleValue" :min="0" :max="100" />
+        <InputNumber v-model:value="inputSingleValue" :min="0" :max="100" />
       </Col>
       <Col :span="18">
         <Slider v-model:value="smallSingleValue" :min="0" :max="1" :step="0.01" />
@@ -187,18 +260,18 @@ const markSingleValue4 = ref(32)
     </Row>
     <h2 class="mt30 mb10">格式化 tooltip</h2>
     <Flex vertical gap="large">
-      <Slider :format-tooltip="formatter" v-model:value="singleValue" />
-      <Slider range :format-tooltip="formatter" v-model:value="doubleValue" />
+      <Slider :format-tooltip="formatter" v-model:value="formatSingleValue" />
+      <Slider range :format-tooltip="formatter" v-model:value="formatDoubleValue" />
     </Flex>
     <h2 class="mt30 mb10">隐藏 tooltip</h2>
     <Flex vertical gap="large">
-      <Slider :tooltip="false" v-model:value="singleValue" />
-      <Slider range :tooltip="false" v-model:value="doubleValue" />
+      <Slider :tooltip="false" v-model:value="hideTooltipSingleValue" />
+      <Slider range :tooltip="false" v-model:value="hideTooltipDoubleValue" />
     </Flex>
     <h2 class="mt30 mb10">自定义样式</h2>
     <Flex vertical gap="large">
-      <Slider :style="customStyle" v-model:value="singleValue" />
-      <Slider :style="rangeCustomStyle" range v-model:value="doubleValue" />
-    </Flex> -->
+      <Slider :style="singleCustomStyle" v-model:value="customStyleSingleValue" />
+      <Slider :style="rangeCustomStyle" range v-model:value="customStyleDoubleValue" :marks="marks" />
+    </Flex>
   </div>
 </template>